% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_2_Achat_neuf_inter.R
\name{achat_neuf_inter}
\alias{achat_neuf_inter}
\title{achat_neuf_inter}
\usage{
achat_neuf_inter(menage, save_intermed_file = F)
}
\arguments{
\item{menage}{A menage dataframe}

\item{save_intermed_file}{Boolean indicating whether the output dataframe should be saved under MatisseFiles$menage_echelle_D2_rd}
}
\value{
A menage dataframe
}
\description{
Function that handles the buying of new houses between the start and the horizon
}
\examples{
achat_neuf_inter(menage, FALSE)

}
