% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_1_alignement_depenses.R
\name{aligne_depenses}
\alias{aligne_depenses}
\title{aligne_depenses}
\usage{
aligne_depenses(menage, save_intermed_file = F)
}
\arguments{
\item{menage}{A menage_echelle dataframe}

\item{save_intermed_file}{Boolean indicating whether the output dataframe should be saved under MatisseFiles$menage_contraintes_rd}
}
\value{
A menage dataframe
}
\description{
aligne_depenses
}
