% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Alignement3ME.R
\name{alignement_3ME}
\alias{alignement_3ME}
\title{alignement_3ME}
\usage{
alignement_3ME(menage_echelle)
}
\arguments{
\item{menage_echelle}{A menage dataframe}
}
\value{
A menage dataframe
}
\description{
Step 5 of Matisse : aligns the emissions on 3ME for fossiles and electricity
}
