% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_util.R
\name{array2list}
\alias{array2list}
\title{array2list}
\usage{
array2list(x)
}
\arguments{
\item{x}{A vector of strings}
}
\value{
A list with each item being an element of the argument vector and values initialized at 0
}
\description{
Array2List is a basic function that transforms a vector of strings into a list with values 0 attached to each element.

    Useful for several functions, notably 'replace_na' which accepts a list as the replace argument
}
\examples{
array2list(c("test1", "test2"))
}
