% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_4_Bascule_Domicile.R
\name{bascule_domicile}
\alias{bascule_domicile}
\title{bascule_domicile}
\usage{
bascule_domicile(menage, save_intermed_file = F)
}
\arguments{
\item{menage}{A menage dataframe}

\item{save_intermed_file}{Boolean indicating whether the output dataframe should be saved under MatisseFiles$menage_echelle_D4_rd}
}
\value{
A menage dataframe
}
\description{
Function that handles the switch of housing category
}
\examples{
bascule_domicile(menage, FALSE)

}
