% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_5_Bascule_Vehicule.R
\name{bascule_efficacite_vehicule}
\alias{bascule_efficacite_vehicule}
\title{bascule_efficacite_vehicule}
\usage{
bascule_efficacite_vehicule(menage, save_intermed_file = F)
}
\arguments{
\item{menage}{A menage dataframe}

\item{save_intermed_file}{Boolean indicating whether the output dataframe should be saved under MatisseFiles$menage_echelle_D5_rd}
}
\value{
A menage dataframe
}
\description{
Function that handles the change of vehicules
}
\examples{
bascule_efficacite_vehicule(menage, FALSE)

}
