% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{calcul_interet_principal}
\alias{calcul_interet_principal}
\title{calcul_interet_principal}
\usage{
calcul_interet_principal(loan, n, year_purchase, horizon, i, pf = 1)
}
\arguments{
\item{loan}{The loaned amount}

\item{n}{The number of payments}

\item{year_purchase}{The year of purchase}

\item{horizon}{The year of our horizon}

\item{i}{The nominal interest rate convertible ic times per year}

\item{pf}{The payment frequency- number of payments per year}
}
\value{
A tibble containing the interest and principal values
}
\description{
Calculates the interest and principal discounted values for a loan of loan euros, over a certain amount of years
Based on the function amort.period -> check it to see the logic behind
}
\examples{
calcul_interet_principal(10000,10,2010,2015,0.01)
}
