% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_1_contraintes.R
\name{calcule_contraintes}
\alias{calcule_contraintes}
\title{calcule_contraintes}
\usage{
calcule_contraintes(menage, save_intermed_file = F)
}
\arguments{
\item{save_intermed_file}{Boolean indicating whether the output dataframe should be saved under MatisseFiles$menage_contraintes_rd}

\item{menage_echelle}{A menage dataframe}
}
\value{
A menage dataframe
}
\description{
This functions calculates the blocking values for the reweighting solver
}
\examples{
calcule_contraintes(menage, FALSE)

}
