% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_Changement_Technique.R
\name{changement_technique}
\alias{changement_technique}
\title{changement_technique}
\usage{
changement_technique(menage)
}
\arguments{
\item{menage}{A menage_echelle dataframe}
}
\value{
A menage_echelle dataframe
}
\description{
Step 3 of Matisse
}
\examples{
changement_technique(menage)

}
