% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_3_estimate_DPE.R
\name{estimate_DPE}
\alias{estimate_DPE}
\title{estimate_DPE}
\usage{
estimate_DPE(menage, save_intermed_file = FALSE)
}
\arguments{
\item{menage}{A menage data frame}

\item{save_intermed_file}{Boolean indicating whether the output dataframe should be saved under MatisseFiles$menage_dpe_rd}
}
\value{
A menage dataframe
}
\description{
This function estimates the DPE class of the menage argument based on the Phebus distribution of DPE and various measures
for each household. Not exported by default.
}
\examples{
estimate_DPE(menage, F)

}
