% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_2_evolution_conso_energie.R
\name{evolution_conso_energie}
\alias{evolution_conso_energie}
\title{evolution_conso_energie}
\usage{
evolution_conso_energie(menage, save_intermed_file = F)
}
\arguments{
\item{menage}{A menage dataframe}

\item{save_intermed_file}{Boolean indicating whether the output dataframe should be saved under MatisseFiles$menage_echelle_2_rd}
}
\value{
A menage dataframe
}
\description{
The evolution of energy consumption
}
\examples{
evolution_conso_energie(menage, FALSE)

}
