% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_1_format_BDF.R
\name{format_BDF}
\alias{format_BDF}
\title{format_BDF}
\usage{
format_BDF(menage, save_intermed_file = FALSE)
}
\arguments{
\item{menage}{A menage data.frame}

\item{save_intermed_file}{Boolean indicating whether the output should be saved under the MatisseFiles$menage_forme_2_rd file}
}
\value{
A menage data.frame
}
\description{
Formats the initial BDF menage file to get 'standardized' variables on all columns. Not exported by default.
}
\examples{
format_BDF(menage, F)

}
