% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_2_format_energie.R
\name{format_energie}
\alias{format_energie}
\title{format_energie}
\usage{
format_energie(menage, save_intermed_file = FALSE)
}
\arguments{
\item{menage}{A menage dataframe}

\item{save_intermed_file}{Boolean indicating whether the output dataframe should be saved under MatisseFiles$menage_forme_3_rd}
}
\value{
A menage dataframe
}
\description{
Adds the energy calculations to a menage data.frame. Not exported by default.
}
\examples{
format_energie(menage, F)

}
