% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_format_menage_data.R
\name{format_menage_data}
\alias{format_menage_data}
\title{format_menage_data}
\usage{
format_menage_data()
}
\value{
A menage_forme dataframe
}
\description{
This is the routing for step A of Matisse run.
}
\examples{
format_menage_data()

}
