% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_3ME_vehicule_stats}
\alias{get_3ME_vehicule_stats}
\title{get_3ME_vehicule_stats}
\usage{
get_3ME_vehicule_stats(year_veh)
}
\arguments{
\item{year_veh}{The year to extract the data at}
}
\value{
A list of various values
}
\description{
Extracts from the MatisseFiles$Threeme_rd file, for the set of params currently used, the different values needed by
Matisse for vehicule simulation
}
\examples{
get_3ME_vehicule_stats(2020)
}
