% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_DPE_stock_year}
\alias{get_DPE_stock_year}
\title{get_DPE_stock_year}
\usage{
get_DPE_stock_year(DPE_year)
}
\arguments{
\item{DPE_year}{The year for the 3ME simulation}
}
\value{
A dataframe contain the class of DPE and the stock for the DPE_year simulation of 3ME
}
\description{
Loads from the file MatisseFiles$sortie_3me_xl, the 3ME standard file, in sheet AMS, the values for the stock of houses for each DPE class
}
\examples{
get_DPE_stock_year(2020)
}
