% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_elast}
\alias{get_elast}
\title{get_elast}
\usage{
get_elast()
}
\value{
A tibble extract of the Excel file
}
\description{
Load the elasticites from the file located in MatisseFiles$elast_xl

Elasticities are typically coming from an econometrics work. Format of the source file is important be careful to respect it
}
\examples{
get_elast()
}
