% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_emissions}
\alias{get_emissions}
\title{get_emissions}
\usage{
get_emissions(menage, FC)
}
\arguments{
\item{menage}{The default format of menage data in Matisse}

\item{FC}{The growth factor, used in every Matisse code}
}
\value{
An array the size of nrow(menage) that contains all the domestic emissions of the households
}
\description{
Calculates the emissions for the the menage data. Uses the file MatisseFiles$coeff_dep_ems_csv as coefficient of emission
between the initial year and the calculated year. Uses coeff_ems_2010_rd as the init year emissions.
}
\examples{
get_emissions(menage, FC)
}
