% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_energie_dom_surf}
\alias{get_energie_dom_surf}
\title{get_energie_dom_surf}
\usage{
get_energie_dom_surf(menage, FC, getFulldf = F)
}
\arguments{
\item{menage}{The default format of menage data in Matisse}

\item{FC}{The growth factor, used in every Matisse code}

\item{getFulldf}{A boolean, default F, that states if you just want to returns the new columns or the full df}
}
\value{
A data frame containing either just the new calculated column of energy consumption or the full menage argument including the new columns
}
\description{
A function that calculates the energy consumption for each household for its home with different values :
total energy consumption, consumption per square-meter.

It read from the file MatisseFiles$prix_class_csv the data for the price for each energy for each household type
}
\examples{
get_energie_dom_surf(menage, FC, F)
}
