% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_gain_energie}
\alias{get_gain_energie}
\title{get_gain_energie}
\usage{
get_gain_energie(conso_moy_dep)
}
\arguments{
\item{conso_moy_dep}{An array of consumption with names being letters from A to G}
}
\value{
A tibble containing the change in percentage in the consumption linked to the change of DPE, depending on the initial and final DPE
}
\description{
Calculates a matrix of energetic gains for a DPE change, according to the conso_moy_dep vector of consumptions
}
\examples{
conso <- 1:7 * 10
names(conso) <- LETTERS[1:7]
get_gain_energie(conso)
}
