% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_list_dep}
\alias{get_list_dep}
\title{get_list_dep}
\usage{
get_list_dep()
}
\value{
An array of strings
}
\description{
Loads a standard vector of 18 strings that contains the default posts of economic behavior for the households in Matisse
}
\examples{
get_list_dep()
}
