% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_list_source_usage}
\alias{get_list_source_usage}
\title{get_list_source_usage}
\usage{
get_list_source_usage()
}
\value{
An array of strings
}
\description{
Gets all the adequat 'source + usage' combinaisons. In particular, only keeps the electric usage for
ElecSpe, Clim and Ecl, which are only electrical
}
\examples{
get_list_source_usage()
}
