% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_list_usages}
\alias{get_list_usages}
\title{get_list_usages}
\usage{
get_list_usages()
}
\value{
An array of strings
}
\description{
Loads a standard vector of the list of possible usages
}
\examples{
get_list_usages()
}
