% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_sources}
\alias{get_sources}
\title{get_sources}
\usage{
get_sources()
}
\value{
An array of string
}
\description{
Loads a standard vector of strings containing the different sources of energy taken into account by Matisse
}
\examples{
get_sources()
}
