% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_summary_dep_table}
\alias{get_summary_dep_table}
\title{get_summary_dep_table}
\usage{
get_summary_dep_table(menage_echelle)
}
\arguments{
\item{menage_echelle}{}
}
\value{
A Sum_horizon_mat
}
\description{
Generates a table that summarises the emissions between menage_echelle and the 3ME data
}
