% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{get_TCO}
\alias{get_TCO}
\title{get_TCO}
\usage{
get_TCO(year_TCO)
}
\value{
A single value, the TCO value for the IMACLIM simulation
}
\description{
Loads the TCO from the file located in MatisseFiles$IMACLIM_3ME_scen_horiz_xl, range C103
}
\examples{
get_TCO()
}
