% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_initialize_Matisse.R
\name{initialize_Matisse}
\alias{initialize_Matisse}
\title{initialize_Matisse}
\usage{
initialize_Matisse()
}
\value{
TRUE
}
\description{
This is the first function to launch Matisse. Please note that it is not best-practice compliant, as it loads the required libraries
directly instead of through proper library management. This requires some work that hasn't yet been scheduled.
}
\examples{
initialize_Matisse()
}
