% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_initialize_Matisse.R
\name{initialize_MatisseFiles}
\alias{initialize_MatisseFiles}
\title{initialize_MatisseFiles}
\usage{
initialize_MatisseFiles()
}
\value{
Returns a global variable named MatisseFiles, which contains all the files links to be used Matisse
}
\description{
initialize_MatisseFiles
}
\examples{
initialize_MatisseFiles()
}
