% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Launcher.R
\name{launch_Matisse}
\alias{launch_Matisse}
\title{launch_Matisse}
\usage{
launch_Matisse()
}
\value{

}
\description{
The main launcher of Matisse. Must have a MatisseParams and MatisseFiles set up to be used
}
