% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_Microsimulation.R
\name{microsimulation}
\alias{microsimulation}
\title{microsimulation}
\usage{
microsimulation(menage)
}
\arguments{
\item{menage}{A menage_echelle dataframe}
}
\value{
A menage_echelle dataframe
}
\description{
Step 2 of Matisse
}
\examples{
microsimulation(menage)

}
