% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_Mise_a_l_echelle.R
\name{mise_a_l_echelle}
\alias{mise_a_l_echelle}
\title{mise_a_l_echelle}
\usage{
mise_a_l_echelle(menage)
}
\arguments{
\item{menage}{A menage_forme dataframe}
}
\value{
A menage_echelle dataframe
}
\description{
Step 1 of Matisse
}
\examples{
mise_a_l_echelle(menage_forme)

}
