% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_1_mise_echelle_revenus.R
\name{mise_echelle_revenus}
\alias{mise_echelle_revenus}
\title{mise_echelle_revenus}
\usage{
mise_echelle_revenus(menage, save_intermed_file = F)
}
\arguments{
\item{menage}{The menage dataframe}

\item{save_intermed_file}{Boolean indicating whether the output dataframe should be saved under MatisseFiles$menage_echelle_B1_rd}
}
\value{
A menage_echelle dataframe
}
\description{
This function puts all the household budget in line to the horizon.
}
\examples{
mise_echelle_revenus(menage_forme, FALSE)

}
