% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_2_ponderation.R
\name{ponderation}
\alias{ponderation}
\title{ponderation}
\usage{
ponderation(menage_contraintes, menage_echelle, save_intermed_file = F)
}
\arguments{
\item{menage_contraintes}{The constraints dataframe}

\item{menage_echelle}{The menage dataframe}
}
\value{
A menage dataframe
}
\description{
Function reweighting
}
\examples{
ponderation(menage_contraintes, menage_echelle)

}
