% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_3_Rehabilitation_inter.R
\name{rehabilitation_inter}
\alias{rehabilitation_inter}
\title{rehabilitation_inter}
\usage{
rehabilitation_inter(menage, save_intermed_file = F)
}
\arguments{
\item{menage}{A menage dataframe}

\item{save_intermed_file}{Boolean indicating whether the output dataframe should be saved under MatisseFiles$menage_echelle_D3_rd}
}
\value{
A menage dataframe
}
\description{
Function that handles the rehabilitation of houses between the start and the horizon
}
\examples{
rehabilitation_inter(menage, FALSE)

}
