% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_Reponderation.R
\name{reponderation}
\alias{reponderation}
\title{ponderation}
\usage{
reponderation(menage_echelle)
}
\arguments{
\item{menage_echelle}{A menage dataframe}
}
\value{
A menage dataframe
}
\description{
ponderation
}
\examples{
reponderation(menage)

}
