% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_2_retrocession_taxe_carbone.R
\name{retrocession_taxe_carbone}
\alias{retrocession_taxe_carbone}
\title{retrocession_taxe_carbone}
\usage{
retrocession_taxe_carbone(menage, save_intermed_file = F)
}
\arguments{
\item{menage}{a menage dataframe}

\item{save_intermed_file}{Boolean indicating whether the output dataframe should be saved under MatisseFiles$menage_echelle_1_rd}
}
\value{
A menage dataframe
}
\description{
The function that calculates the retrocession of carbon taxe to households
}
\examples{
retrocession_taxe_carbone(menage, FALSE)

}
