% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{select_bascule}
\alias{select_bascule}
\title{select_bascule}
\usage{
select_bascule(
  sub_men_ech,
  Sum_horizon_Mat,
  scenario_classement_bascule,
  redistribution,
  ener
)
}
\arguments{
\item{redistribution}{}
}
\value{
A list of data with the selection of bascules
}
\description{
select_bascule
}
