% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_1_simulation_depenses.R
\name{simulation_depenses}
\alias{simulation_depenses}
\title{simulation_depenses}
\usage{
simulation_depenses(menage, save_intermed_file = F)
}
\arguments{
\item{menage}{The menage dataframe}

\item{save_intermed_file}{Boolean indicating whether the output dataframe should be saved under MatisseFiles$menage_echelle_C1_rd}
}
\value{
A menage dataframe
}
\description{
This function simulates the budgets change
}
\examples{
simulation_depenses(menage, FALSE)

}
