% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_functions_Matisse.R
\name{ventilate_solde}
\alias{ventilate_solde}
\title{ventilate_solde}
\usage{
ventilate_solde(menage, solde, FC, step_ventil)
}
\arguments{
\item{menage}{Standard menage data.frame for Matisse}

\item{solde}{An array containing the value to be distributed for each household}

\item{FC}{The growth factor to be used for each spending post}

\item{step_ventil}{The type of ventilation : VE or REHAB}
}
\value{
Returns a modified menage data.frame
}
\description{
Ventilate the argument solde across the different spending posts. Based on the standard menage data.frame for Matisse,
solde an array of the solde to be distributed for each menage, FC the growth factor.
}
\examples{
ventilate_solde(menage, solde, FC, "VE")
}
